+++ completion/_balena
@@ -8,10 +8,9 @@ _balena() {
   local context state line curcontext="$curcontext"
 
   # Valid top-level completions
-  main_commands=( apps build deploy envs fleets join keys leave login logout logs note orgs preload push releases scan settings ssh support tags tunnel version whoami api-key app app config device device devices env fleet fleet internal key key local os release release tag util )
+  main_commands=( build deploy envs fleets join keys leave login logout logs note orgs preload push releases scan settings ssh support tags tunnel version whoami api-key config device device devices env fleet fleet internal key key local os release release tag util )
   # Sub-completions
   api_key_cmds=( generate )
-  app_cmds=( create purge rename restart rm )
   config_cmds=( generate inject read reconfigure write )
   device_cmds=( deactivate identify init local-mode move os-update public-url purge reboot register rename restart rm shutdown )
   devices_cmds=( supported )
@@ -43,9 +42,6 @@ _balena_sec_cmds() {
   case $line[1] in
       "api-key")
         _describe -t api_key_cmds 'api-key_cmd' api_key_cmds "$@" && ret=0
-      ;;
-      "app")
-        _describe -t app_cmds 'app_cmd' app_cmds "$@" && ret=0
       ;;
       "config")
         _describe -t config_cmds 'config_cmd' config_cmds "$@" && ret=0
