+++ completion/balena-completion.bash
@@ -7,10 +7,9 @@ _balena_complete()
   local cur prev
 
   # Valid top-level completions
-  main_commands="apps build deploy envs fleets join keys leave login logout logs note orgs preload push releases scan settings ssh support tags tunnel version whoami api-key app app config device device devices env fleet fleet internal key key local os release release tag util"
+  main_commands="build deploy envs fleets join keys leave login logout logs note orgs preload push releases scan settings ssh support tags tunnel version whoami api-key config device device devices env fleet fleet internal key key local os release release tag util"
   # Sub-completions
   api_key_cmds="generate"
-  app_cmds="create purge rename restart rm"
   config_cmds="generate inject read reconfigure write"
   device_cmds="deactivate identify init local-mode move os-update public-url purge reboot register rename restart rm shutdown"
   devices_cmds="supported"
@@ -37,9 +36,6 @@ _balena_complete()
     case "$prev" in
       api-key)
         COMPREPLY=( $(compgen -W "$api_key_cmds" -- $cur) )
-        ;;
-      app)
-        COMPREPLY=( $(compgen -W "$app_cmds" -- $cur) )
         ;;
       config)
         COMPREPLY=( $(compgen -W "$config_cmds" -- $cur) )
