+++ lib/commands/fleet/restart.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2021 Balena Ltd.
+ * Copyright 2016-2020 Balena Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +15,56 @@
  * limitations under the License.
  */
 
-import { FleetRestartCmd } from '../app/restart';
+import type { flags } from '@oclif/command';
 
-export default FleetRestartCmd;
+import Command from '../../command';
+import * as cf from '../../utils/common-flags';
+import * as ca from '../../utils/common-args';
+import { getBalenaSdk, stripIndent } from '../../utils/lazy';
+import { applicationIdInfo } from '../../utils/messages';
+
+interface FlagsDef {
+	help: void;
+}
+
+interface ArgsDef {
+	fleet: string;
+}
+
+export default class FleetRestartCmd extends Command {
+	public static description = stripIndent`
+		Restart a fleet.
+
+		Restart all devices belonging to a fleet.
+
+		${applicationIdInfo.split('\n').join('\n\t\t')}
+	`;
+
+	public static examples = [
+		'$ balena fleet restart MyFleet',
+		'$ balena fleet restart myorg/myfleet',
+	];
+
+	public static args = [ca.fleetRequired];
+
+	public static usage = 'fleet restart <fleet>';
+
+	public static flags: flags.Input<FlagsDef> = {
+		help: cf.help,
+	};
+
+	public static authenticated = true;
+
+	public async run() {
+		const { args: params } = this.parse<FlagsDef, ArgsDef>(FleetRestartCmd);
+
+		const { getApplication } = await import('../../utils/sdk');
+
+		const balena = getBalenaSdk();
+
+		// Disambiguate application (if is a number, it could either be an ID or a numerical name)
+		const application = await getApplication(balena, params.fleet);
+
+		await balena.models.application.restart(application.id);
+	}
+}
